fix: change upload to a temporary folder

+++ controllers/profiles.js
@@ -332,9 +332,10 @@ profilesRouter.post("/upload", upload.single("file"), async (req, res) => {
     // Hand the file path off to your dedicated service
     const finalStats = await processCSV(req.file.path);
 
-    // Clean up the temp file
-    fs.unlinkSync(req.file.path);
-
+    // Clean up is still important!
+    if (fs.existsSync(req.file.path)) {
+      fs.unlinkSync(req.file.path);
+    }
     // Return the required JSON
     return res.status(200).json(finalStats);
   } catch (error) {
